Talk about host specific/independent instead of mutability
Cover my action item from this week's OCI call.

Also moved State json out of config.go since its not really part of
the config.json file/config.
